requirements.txt
gensim==3.6.0
flask==1.0.2
wget==3.2
ufal.udpipe==1.2.0.1
nltk==3.3
gunicorn==19.9.0